anniversaireback/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.2
python_bcrypt==0.3.2
